+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # gerardroche/sublime_color_scheme_unit
 
-## Unreleased
+## 0.7.0
 
 * Added: improved error handling; prints error information in test results
 * Added: allow tests to begin with comments and tags e.g. `<?php // COLOR TEST ...`